pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hf-ms-transfer"
version = "0.4.0"
authors = [
  { name="wa008", email="zphu1024@gmail.com" },
]
description = "A tool to transfer models/datasets from Hugging Face to ModelScope"
readme = "README.md"
requires-python = ">=3.6"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "huggingface_hub",
    "modelscope",
    "python-dotenv",
]

[project.urls]
"Homepage" = "https://github.com/wa008/hf-ms-transfer"
"Bug Tracker" = "https://github.com/wa008/hf-ms-transfer/issues"

[project.scripts]
"hf-ms-transfer" = "hf_ms_transfer.__main__:main"
